+++ resources/views/leveranciers/table.blade.php
@@ -20,19 +20,34 @@
             <td>{{ $leveranciers->directeur }}</td>
             <td>{{ $leveranciers->telefoonnummer }}</td>
             <td>{{ $leveranciers->website }}</td>
+            @if (request()->path() !== 'restore-leveranciers')
+
                 <td width="120">
                     {!! Form::open(['route' => ['leveranciers.destroy', $leveranciers->id], 'method' => 'delete']) !!}
                     <div class='btn-group'>
-                        <a href="{{ route('leveranciers.show', [$leveranciers->id]) }}" class='btn btn-default btn-xs'>
+                        <a href="{{ route('leveranciers.show', [$leveranciers->id]) }}" class='btn btn-default btn-s'>
                             <i class="far fa-eye"></i>
                         </a>
-                        <a href="{{ route('leveranciers.edit', [$leveranciers->id]) }}" class='btn btn-default btn-xs'>
+                        <a href="{{ route('leveranciers.edit', [$leveranciers->id]) }}" class='btn btn-default btn-s'>
                             <i class="far fa-edit"></i>
                         </a>
-                        {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                        {!! Form::button('<i class="far fa-trash-alt"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-s', 'onclick' => "return confirm('Are you sure?')"]) !!}
                     </div>
                     {!! Form::close() !!}
                 </td>
+
+                @else
+                <td width="120">
+                    {!! Form::open(['route' => ['leveranciers.restoreBack', $leveranciers->id], 'method' => 'post']) !!}
+                    <div class='btn-group'>
+                        <a href="{{ route('leveranciers.restoreShow', [$leveranciers->id]) }}" class='btn btn-default btn-s'>
+                            <i class="far fa-eye"></i>
+                        </a>
+                        {!! Form::button('<i class="fas fa-trash-restore"></i>', ['type' => 'submit', 'class' => 'btn btn-warning btn-s', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                    </div>
+                    {!! Form::close() !!}
+                </td>
+                @endif
             </tr>
         @endforeach
         </tbody>
